ci: Run go tests in github action
This runs the make targets:

- make test-ci
- make test-api
- make test-sdk

The go tests in the main module are split by package and run in parallel
in a matrix build with GOMAXPROCS=1. This is done primarily to work
around an issue with hitting the ulimit on nofiles on github runners.
When running at a higher GOMAXPROCS tests that use vault instances in
docker containers start to intermitently fail with errors from vault,
usually either a 400, 403, or TLS error. It is believed this is due to
reaching the file limit.

At this time, it appears that the github public runners have a hard
file limit of 65536. In CircleCI, the hard limit is 524288. This
includes a step in the workflow to output the limits. If the limit is
increased we can try increasing GOMAXPROCS.

Co-authored-by: Sarah Thompson <sthompson@hashicorp.com>
Co-authored-by: Daniel Kimsey <daniel.kimsey@hashicorp.com>